refactor: move code to shared folder
